[products] Updates latest, adds latestReleaseDate
- This information is getting picked up from the release-data repository
- This should (hopefully) get auto-updated along with every dependabot
  PR that is raised in the repo by itself. We should still review those
  I think to make sure there aren't any bugs
- Some dates are off by 1 or 2 days - mainly because our earlier
  information comes from blog posts and news pages, while this is coming
  from the tag information. Adding timezone confusion (we calculate
  dates on UTC for now) makes the drift slightly larger in some cases -
  I think it's acceptable.

The update script uses ruamel, which treats newlines in lists as
belonging to the previous element, which causes some weird newlines.
That's not a big issue as well.

The only quirk with this is that ruamel needs a clean newline after the
`---` in the markdown file. Will add a test for that somewhere.